requirements.txt
fastapi
uvicorn
pydantic
pdfplumber
tesserocr
python-multipart
tqdm
requests
Pillow
python-dotenv
llama-cpp-python
pytest
